Move tests from BasicIT in in order not to repeat them for OverlayIT
