+++ vaadin-login-flow-integration-tests/src/main/java/com/vaadin/flow/component/login/examples/OverlaySelfAttachedView.java
@@ -1,15 +1,15 @@
 package com.vaadin.flow.component.login.examples;
 
+import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.NativeButton;
 import com.vaadin.flow.component.login.LoginOverlay;
 import com.vaadin.flow.router.Route;
 
 @Route
-public class OverlaySelfAttachedView extends AbstractView {
-
-    private final LoginOverlay login = new LoginOverlay();
+public class OverlaySelfAttachedView extends Div {
 
     public OverlaySelfAttachedView() {
+        LoginOverlay login = new LoginOverlay();
         login.addLoginListener(e -> login.close());
         NativeButton button = new NativeButton("open");
         button.setId("open");
