+++ vaadin-login-flow-integration-tests/src/test/java/com/vaadin/flow/component/login/test/BasicIT.java
@@ -40,41 +40,6 @@ public abstract class BasicIT extends AbstractParallelTest {
     }
 
     @Test
-    public void enterKeyLogin() {
-        if (BrowserUtil.isEdge(getDesiredCapabilities())) {
-            skipTest("Skip for Edge due to the sendKeys usage");
-        }
-        LoginElement login = $(LoginElement.class).waitForFirst();
-        checkSuccessfulLogin(login, () -> {
-            login.focus();
-            login.sendKeys(Keys.ENTER);
-        });
-    }
-    @Test
-    public void passwordEnterKeyLogin() {
-        if (BrowserUtil.isEdge(getDesiredCapabilities())) {
-            skipTest("Skip for Edge due to the sendKeys usage");
-        }
-        LoginElement login = $(LoginElement.class).waitForFirst();
-        checkSuccessfulLogin(login, () -> {
-            login.getPasswordField().focus();
-            login.sendKeys(Keys.ENTER);
-        });
-    }
-
-    @Test
-    public void usernameEnterKeyLogin() {
-        if (BrowserUtil.isEdge(getDesiredCapabilities())) {
-            skipTest("Skip for Edge due to the sendKeys usage");
-        }
-        LoginElement login = $(LoginElement.class).waitForFirst();
-        checkSuccessfulLogin(login, () -> {
-            login.getUsernameField().focus();
-            login.sendKeys(Keys.ENTER);
-        });
-    }
-
-    @Test
     public void login() {
         AbstractLoginElement login = getLogin();
         checkSuccessfulLogin(login, () -> login.submit());
